refactor: consolidate config under config/
