+++ frontend/hooks/use-location.ts
@@ -2,7 +2,7 @@
 
 import { useEffect, useMemo, useRef, useState } from "react";
 
-import { GOOGLE_MAPS_API_KEY } from "@lib/config";
+import { GOOGLE_MAPS_API_KEY } from "@config/env";
 
 declare global {
   interface Window {
