+++ frontend/lib/supabase-browser.ts
@@ -1,7 +1,7 @@
 import { createBrowserClient } from "@supabase/ssr";
 import type { SupabaseClient } from "@supabase/supabase-js";
 
-import { SUPABASE_ANON_KEY, SUPABASE_URL } from "@lib/config";
+import { SUPABASE_ANON_KEY, SUPABASE_URL } from "@config/env";
 
 let browserClient: SupabaseClient | undefined;
